Rework the client and command line tool
Code has not been ready to receive update that would allow us to
upload assets to release, so we changed it a bit. Main difference is
explicit URL passing to all API calls, which makes it much easier to
reason about the call being made.

Another additions is use of keyword arguments in all wrappers for GET,
POST and DELETE calls that makes it possible to pass arbitrary
additional arguments to requests.